+++ js_modules/dagit/packages/core/src/instance/InstanceRoot.tsx
@@ -2,9 +2,11 @@ import {MainContent} from '@dagster-io/ui';
 import * as React from 'react';
 import {Redirect, Route, Switch, useLocation} from 'react-router-dom';
 
+import {useFeatureFlags} from '../app/Flags';
 import {AssetsCatalogRoot} from '../assets/AssetsCatalogRoot';
 import {AssetsGroupsGlobalGraphRoot} from '../assets/AssetsGroupsGlobalGraphRoot';
 import {RunRoot} from '../runs/RunRoot';
+import {RunTimelineRoot} from '../runs/RunTimelineRoot';
 import {RunsRoot} from '../runs/RunsRoot';
 import {ScheduledRunListRoot} from '../runs/ScheduledRunListRoot';
 import {SnapshotRoot} from '../snapshots/SnapshotRoot';
@@ -13,6 +15,7 @@ import {InstanceStatusRoot} from './InstanceStatusRoot';
 
 export const InstanceRoot = () => {
   const {pathname} = useLocation();
+  const {flagNewWorkspace} = useFeatureFlags();
   const main = React.useRef<HTMLDivElement>(null);
 
   React.useEffect(() => {
@@ -31,6 +34,11 @@ export const InstanceRoot = () => {
         <Route path="/instance/runs" exact>
           <RunsRoot />
         </Route>
+        {flagNewWorkspace ? (
+          <Route path="/instance/runs/timeline" exact>
+            <RunTimelineRoot />
+          </Route>
+        ) : null}
         <Route path="/instance/runs/scheduled" exact>
           <ScheduledRunListRoot />
         </Route>
