+++ js_modules/dagit/packages/core/src/instance/JobTable.tsx
@@ -4,9 +4,9 @@ import * as React from 'react';
 import {ScheduleOrSensorTag} from '../nav/ScheduleOrSensorTag';
 import {LegacyPipelineTag} from '../pipelines/LegacyPipelineTag';
 import {PipelineReference} from '../pipelines/PipelineReference';
-import {makeJobKey} from '../runs/QueryfulRunTimeline';
 import {RunStatusPezList} from '../runs/RunStatusPez';
 import {RunTimeFragment} from '../runs/types/RunTimeFragment';
+import {makeJobKey} from '../runs/useRunsForTimeline';
 import {ScheduleSwitchFragment} from '../schedules/types/ScheduleSwitchFragment';
 import {SensorSwitchFragment} from '../sensors/types/SensorSwitchFragment';
 import {repoAddressAsString} from '../workspace/repoAddressAsString';
