+++ js_modules/dagit/packages/core/src/workspace/WorkspaceJobsRoot.tsx
@@ -18,10 +18,10 @@ import {isHiddenAssetGroupJob} from '../asset-graph/Utils';
 import {JobItemWithRuns, JobTable, ScheduleFragment} from '../instance/JobTable';
 import {SCHEDULE_FUTURE_TICKS_FRAGMENT} from '../instance/NextTick';
 import {RepoFilterButton} from '../instance/RepoFilterButton';
-import {makeJobKey} from '../runs/QueryfulRunTimeline';
 import {RepoSectionHeader} from '../runs/RepoSectionHeader';
 import {RUN_TIME_FRAGMENT} from '../runs/RunUtils';
 import {RunTimeFragment} from '../runs/types/RunTimeFragment';
+import {makeJobKey} from '../runs/useRunsForTimeline';
 import {SCHEDULE_SWITCH_FRAGMENT} from '../schedules/ScheduleSwitch';
 import {SENSOR_SWITCH_FRAGMENT} from '../sensors/SensorSwitch';
 import {SensorSwitchFragment} from '../sensors/types/SensorSwitchFragment';
